Fix broken DOM reference
The build was failing because we had the `onload` in the loaded JS file,
and when `karma` tried running that code outside of a browser, there
was no DOM and the `element` lookup failed.

I'm not sure this is a great solution, but this moves the `onload`
code into the HTML since we're guaranteed to have a DOM there.