Encourage them to slide the cake with their stories
The way we lay out the tasks in `LABTASKS.md` tends to isolate the
front end work from the backend work. We want them to create stories
that do a better job of slicing the cake than a straightforward reading
would suggest. I added some test to try to help with that.

This helps with #31, but I'm not sure it finishes it. We also want this
reflected in the rubric, and I haven't had time to see if we do that
or not.