Proofread and update lab write-up
The one really major change is a complete re-write of the tasks so that
everything now slices the cake. I have them creating an epic for each
feature, with (probably) two stories per epic (one for the server-side
API code, and one for the client-side support).

There are also a variety of edits that hopefully improve the write-up. Things
I tried to improve:

   * Specifying the JDK location in IntelliJ
   * Importing a project into IntelliJ
   * What it means to run the server
   * How to run JUnit tests
   * How to set up the Travis badge
   * Fix some broken links
   * Clarify the questions in `LABTASKS.md`
   * Link to some info on how create Epis in ZenHub

 I removed the material about creating a new Gradle task, as that
 didn't actually work and it certainly wasn't very clear.

 We changed from having them create an `ANSWERS.md` file in their repo
 to creating a GDoc with their answers and turning that in via Canvas.

Closes #14
Closes #27
Closes #29
Closes #31
Closes #33
Closes #39
Closes #40
Closes #41
Closes #42
Closes #48