Isolate GET request logic into util.js
Per #53, this commit takes the HttpClient() function definition previously in users.js and moves it to its own file (util.js) to provide a better example of code organization.

Once complete, this improvement will also take care of #52.

As of this commit, util.js is not actually being imported into users.js. I was having some trouble finding a way of importing that actually works in vanilla JS.